HouseAffordability
Not finished. However, got functionality for when [Constraints: Down Payment is a percentage only and when  property tax, Hoa, and insurance are either percentages/dollar value.]
Also no test cases yet. 
This one formula allowed for functionality for Conventional loan, and 10% up to 50%. 
FHA, VA will be other calculations. Unfortunately I do not have them yet.
I have the formula for when down payment is in cash format but have not put it in yet, ran out of time this week need to do other homework lol.
Problem, when interest rate is 0, looks like I need to find a new formula cause it breaks. 
Very close.