House affordability
Hey guys, can you please try to tell me if my calculator works/see if it crashes for a certain input or wrong output for something.
It looks messy for now, but I think I finally finished everything that is possible. 
I was actually able to implement all options, except of course FHA loan, a pop up will appear for that.
Also when cash is chosen for downpayment, it will give you a pop up when the user inputs like 100,000 grand. Basically it will output wrong numbers when down payment is over 20%.
Next up is seeing for test cases that crash, and then designing this calculator then my rent calculator as well.